[refactor] #238 UserService CQRS 패턴 적용
